+++ packages/terra-date-picker/src/react-datepicker/calendar.jsx
@@ -1,4 +1,4 @@
-import { FormattedMessage, intlShape } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 import * as KeyCode from 'keycode-js';
 import YearDropdown from './year_dropdown'
 import MonthDropdown from './month_dropdown'
@@ -105,6 +105,11 @@ export default class Calendar extends React.Component {
      */
     includeDates: PropTypes.array,
     /**
+     * Timezone value to indicate in which timezone the date-time component is rendered.
+     * The value provided should be a valid [timezone](https://en.wikipedia.org/wiki/List_of_tz_database_time_zones) string, else will default to browser/local timezone.
+     */
+    initialTimeZone: PropTypes.string,
+    /**
      * Prop to show inline version of date picker component.
      */
     inline: PropTypes.bool,
@@ -112,7 +117,7 @@ export default class Calendar extends React.Component {
      * @private
      * Internationalization object with translation APIs. Provided by `injectIntl`.
      */
-    intl: intlShape,
+    intl: PropTypes.shape({ formatMessage: PropTypes.func }),
     /**
      * Name of locale data for different international formatting.
      */
@@ -219,10 +224,6 @@ export default class Calendar extends React.Component {
      * Prop to show short names of weekdays .
      */
     useWeekdaysShort: PropTypes.bool,
-    /**
-     * Difference between utc and local time.
-     */
-    utcOffset: PropTypes.number,
     /**
      * Label value for weeks on date picker.
      */
@@ -387,10 +388,10 @@ export default class Calendar extends React.Component {
   }
 
   getDateInView = () => {
-    const { preSelection, selected, openToDate, utcOffset } = this.props
+    const { preSelection, selected, openToDate, initialTimeZone } = this.props
     const minDate = getEffectiveMinDate(this.props)
     const maxDate = getEffectiveMaxDate(this.props)
-    const current = now(utcOffset)
+    const current = now(initialTimeZone)
     const initialDate = openToDate || selected || preSelection
     if (initialDate) {
       return initialDate
@@ -606,7 +607,7 @@ export default class Calendar extends React.Component {
       return
     }
 
-    const today = getStartOfDate(now(this.props.utcOffset));
+    const today = getStartOfDate(now(this.props.initialTimeZone));
     return (
       <button
         className={cx('react-datepicker-today-button')}
@@ -683,10 +684,10 @@ export default class Calendar extends React.Component {
             startDate={this.props.startDate}
             endDate={this.props.endDate}
             peekNextMonth={this.props.peekNextMonth}
-            utcOffset={this.props.utcOffset}
             handleCalendarKeyDown={this.props.handleCalendarKeyDown}
             locale={this.props.locale}
-            intl={this.props.intl} />
+            intl={this.props.intl}
+            initialTimeZone={this.props.initialTimeZone} />
           <div className={cx('react-datepicker-header')}>
             {this.renderCurrentMonth(monthDate)}
             <div className={cx('react-datepicker-header-controls')}>