+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerDST.test.jsx
@@ -4,6 +4,7 @@ import DateTimePickerExampleTemplate from '../../common/DateTimePickerExampleTem
 const DateTimePickerExample = () => (
   <DateTimePickerExampleTemplate
     value="2017-11-05T01:30:00"
+    initialTimeZone="America/Chicago"
   />
 );