+++ packages/terra-application-layout/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to support a peer dependency of react-intl v2-v5
+  * Removed `intlShape` imports
+
+* Fixed
+ * Fixed broken links in documentation.
+
+## 5.47.0 - (December 8, 2020)
+
+* Changed
+  * Minor dependency version bump
+
 ## 5.46.0 - (November 24, 2020)
 
 * Changed