+++ packages/terra-application-utility/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to support a peer dependency of react-intl v2-v5
+  * Removed `intlShape` imports
+
+* Fixed
+ * Fixed broken links in documentation.
+
 ## 2.40.0 - (November 24, 2020)
 
 * Changed
@@ -20,7 +27,7 @@
 
 * Changed
   * Updated Changelog format
-  
+
 ## 2.37.0 - (August 4, 2020)
 
 * Changed