+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+## 4.54.0 - (December 8, 2020)
+
+* Added
+  * Added initialTimeZone a private prop that gets passed down from terra-date-time-picker.
+
+* Changed
+  * Changed 'moment' dependency to 'moment-timezone'
+  * Updated to support a peer dependency of react-intl v2-v5
+
 ## 4.53.0 - (November 24, 2020)
 
 * Changed