+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -2,11 +2,25 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to support a peer dependency of react-intl v2-v5
+  * Updated to use `injectIntl` instead of `FormattedMessage` for react-intl v2-v5 compatibility
+
+* Fixed
+ * Fixed broken links in documentation.
+
+## 4.54.0 - (December 8, 2020)
+
+* Added
+  * Added `initialTimeZone` prop which takes a valid timezone string. This enables date-time-picker set to a particular timezone. The default timezone is set to the local/browser timezone.
+
+* Fixed
+  * Fixed flaky test failure due to loading unintended test example.
+
 ## 4.53.0 - (November 24, 2020)
 
 * Changed
   * Update wdio tests to remove deprecated Terra.it helpers
-
 
 ## 4.52.0 - (November 10, 2020)
 
@@ -38,8 +52,6 @@
 * Changed
   * `onClickOutside` prop has been deprecated, will resolve to `onRequestClose`
   * Updated wdio snapshots due to changes in date-picker and time-input
-
-* Changed
   * Update `timeVariant` prop description to include `es` in locales supporting only 24 hour format.
 
 ## 4.46.0 - (September 10, 2020)