+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Fixed
+ * Fixed broken links in documentation.
+
+## 4.6.0 - (December 8, 2020)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.5.0 - (November 24, 2020)
 
 * Changed
@@ -10,8 +18,9 @@
 ## 4.4.0 - (November 10, 2020)
 
 * Changed
+  * Updated to support a peer dependency of react-intl v2-v5
   * Updated jest  due to changes in abstract modal.
-  
+
 ## 4.3.0 - (October 20, 2020)
 
 * Changed