+++ packages/terra-popup/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to support a peer dependency of react-intl v2-v5
+  * Updated FormattedMessage child function to join if given an array.
+
+## 6.45.0 - (December 8, 2020)
+
+* Changed
+  * Minor dependency version bump
+
 ## 6.44.0 - (November 24, 2020)
 
 * Changed