+++ packages/terra-slide-panel/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Fixed
+ * Fixed broken links in documentation.
+ 
+## 3.30.0 - (December 8, 2020)
+
+* Changed
+  * Update wdio tests to remove deprecated Terra.it helpers
+
 ## 3.29.0 - (August 4, 2020)
 
 * Changed