+++ packages/terra-tabs/CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## Unreleased
 
+* Fixed
+ * Fixed broken links in documentation.
+
+## 6.47.0 - (December 8, 2020)
+
 * Changed
+  * Updated to support a peer dependency of react-intl v2-v5
+  * Removed `intlShape` import
   * Update wdio tests to remove deprecated Terra.it helpers
  
 ## 6.46.0 - (November 24, 2020)