+++ packages/terra-time-input/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+* Fixed
+ * Fixed broken links in documentation.
+
+## 4.36.0 - (December 8, 2020)
+
+* Changed
+  * Updated to support a peer dependency of react-intl v2-v5
+  * Update wdio tests to remove deprecated Terra.it helpers
+
 ## 4.35.0 - (November 10, 2020)
 
 * Fixed